[Bugfix] Selecting table then pressing delete removes it

+++ assets/src/components/editing/editor/handlers/deleteblock.ts
@@ -1,9 +1,11 @@
 import { KeyboardEvent } from 'react';
-import { Range, Editor as SlateEditor, Transforms } from 'slate';
+import { Editor, Path, Range, Editor as SlateEditor, Transforms } from 'slate';
 
 const blocksToDelete: string[] = ['table'];
+const internalBlocks: string[] = ['td', 'th'];
 
 const matchBlockToDelete = (node: any) => blocksToDelete.includes(node.type);
+const matchInternalBlock = (node: any) => internalBlocks.includes(node.type);
 
 const deleteOrBackspaceBlock = (
   editor: SlateEditor,
@@ -11,11 +13,29 @@ const deleteOrBackspaceBlock = (
   direction: 'before' | 'after',
 ) => {
   try {
-    if (
-      (e.key === 'Delete' || e.key === 'Backspace') &&
-      editor.selection &&
-      Range.isCollapsed(editor.selection)
-    ) {
+    if (editor.selection && !Range.isCollapsed(editor.selection)) {
+      // When the range isn't collapsed, we want to delete the overall table if the selection spans more than one cell.
+      const [start, end] = Range.edges(editor.selection);
+      const startItem = SlateEditor.above(editor, {
+        at: start,
+        match: matchInternalBlock,
+      });
+      const endItem = SlateEditor.above(editor, {
+        at: end,
+        match: matchInternalBlock,
+      });
+
+      if (startItem && endItem && Path.compare(startItem[1], endItem[1]) !== 0) {
+        const deletableItem = SlateEditor.above(editor, {
+          at: start,
+          match: matchBlockToDelete,
+        });
+        if (deletableItem) {
+          Transforms.removeNodes(editor, { at: deletableItem[1] });
+          e.preventDefault();
+        }
+      }
+    } else if (editor.selection && Range.isCollapsed(editor.selection)) {
       const [currentCursorPosition] = Range.edges(editor.selection);
 
       const nextCursorPosition =
@@ -48,9 +68,13 @@ const deleteOrBackspaceBlock = (
 };
 
 export const deleteBlockKeyDown = (editor: SlateEditor, e: KeyboardEvent) => {
-  deleteOrBackspaceBlock(editor, e, 'after');
+  if (e.key === 'Delete') {
+    deleteOrBackspaceBlock(editor, e, 'after');
+  }
 };
 
 export const backspaceBlockKeyDown = (editor: SlateEditor, e: KeyboardEvent) => {
-  deleteOrBackspaceBlock(editor, e, 'before');
+  if (e.key === 'Backspace') {
+    deleteOrBackspaceBlock(editor, e, 'before');
+  }
 };
